slave/requirements.txt
aio_pika==9.5.8
asyncpg==0.31.0
py3_validate_email==1.0.5.post2
